[Refactor] Convert reagent modules to packages
