+++ reagent/models/__init__.py
@@ -1,6 +1,6 @@
 """Re-export commonly used models for convenience."""
 
-from .protocols import ParsedArtifact, NormalizedModel, ValidationMessage
+from ..protocols import ParsedArtifact, NormalizedModel, ValidationMessage
 
 __all__ = [
     "ParsedArtifact",
